split the table.rs file into multiple submodules
