+++ crates/nu-command/tests/commands/table/mod.rs
@@ -1,0 +1,10 @@
+mod abbreviation;
+mod collapse_expand;
+mod display;
+mod footer;
+mod indexes;
+mod large_outputs;
+mod padding;
+mod streams_metadata;
+mod themes;
+mod width_priority;
